requirements.txt
faiss-gpu==1.6.3
joblib>=1.2.0
matplotlib==3.2.1
numba==0.49.1
numpy>=1.22
Pillow>=9.0.1
scikit-learn==0.23.0
scipy==1.4.1
sklearn==0.0
# torch==1.5.1+cu101
# torchvision==0.6.1+cu101
tqdm==4.46.0
